Fall back to sorting the locs if the names are not enough
Previously, the implementation would call showRaw on the entire symbol,
which meant that comparing two symbols would be accidentally quadratic.

Comparing two symbols only requires looking at the symbol's name, not
its children. Names are unique, but their printings are not--names have
IDs in them that are not always stable from one run to another, and are
also brittle in the sense that slight changes in the input program can
cause almost all IDs in a program to change.

So when we print Symbols, we strip out the IDs and replace them all with
`$CENSORED`, which means that names that are nominally distinct are
actually not printed distinctly.

The only time we do this for names is for `<static-init>` names. Those
IDs also have another way to sort them: by the file path that that
`<static-init>` method lives in. So we compare by these instead.